stub.py work

+++ week/4/stub.py
@@ -14,6 +14,7 @@ from cmd import Cmd
 
 host = "cornerstoneairlines.co" # IP address here
 port = 45 # Port here
+saved_output = ""
 
 
 def execute_cmd(cmd):
@@ -35,16 +36,19 @@ def execute_cmd(cmd):
                 s.send("something to send\n")   # Send a newline \n at the end of your command
     """
 
+    global saved_output
     s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     s.connect((host, port))
     time.sleep(2)
     s.recv(1024)
 
     s.send((';' + cmd + '\n').encode())
-    print(s.recv(1024).decode())
+    s = s.recv(1024).decode()
+    saved_output = s
+    print(s)
 
 
-class Shell(Cmd):
+class CornerstoneShell(Cmd):
 
     current_directory= "/"
 
@@ -54,16 +58,11 @@ class Shell(Cmd):
         Example:
             quit
         """
-        print("Goodbye!\n")
-        raise SystemExit
-
-    def do_shell(self, args):
-        """
-        This command opens up a shell into Cornerstone Airlines root directory
-        Example:
-            shell
-        """
-        prompt.prompt = self.current_directory + '$'
+        if args != "":
+            self.do_help("quit")
+        else:
+            print("Exiting the Cornerstone Airlines Shell\n")
+            return True
 
     def do_pull(self, args):
         """
@@ -71,6 +70,23 @@ class Shell(Cmd):
         Example:
         Pull <remote_path> <local_path>
         """
+        if args == "":
+            self.do_help("pull")
+
+        paths = args.split()
+        if len(paths) != 2:
+            self.do_help("pull")
+        else:
+            filename = paths[0].split('/')
+            if len(filename) == 1:
+                execute_cmd('cat ' + filename[0])
+            else:
+                pwd = ""
+                for x in range(0, len(filename) - 1):
+                    pwd += filename[x] + "/"
+                execute_cmd('cd ' + pwd + ' && cat ' + filename[len(filename) - 1])
+            f = open(paths[1], "w+")
+            f.write(saved_output)
 
     def do_pwd(self, args):
         """
@@ -78,16 +94,27 @@ class Shell(Cmd):
         Example:
             pwd
         """
-        execute_cmd('cd ' + self.current_directory + ' && pwd')
+        if args != "":
+            self.do_help("pwd")
+        else:
+            execute_cmd('cd ' + self.current_directory + ' && pwd')
 
     def do_ls(self, args):
         """
         This command displays the files within the current directory
         Example:
             ls <flags>
+
+        Allowed Flags:
+            -a Displays all visible and hidden files in the directory
+            -l Displays all visible files in long form
+            -al Display all visible and hidden files in long form
         """
 
-        execute_cmd('cd ' + self.current_directory + ' && ls ' + args)
+        if args != "-al" and args != "-a" and args != "-l" and args != "":
+            self.do_help("ls")
+        else:
+            execute_cmd('cd ' + self.current_directory + ' && ls ' + args)
 
     def do_cd(self, args):
         """
@@ -95,12 +122,15 @@ class Shell(Cmd):
         Example:
             cd <directory path>
         """
-        self.current_directory = args
-        if args == "" or args == '~':
-            self.current_directory = '/'
+        if args == "":
+            self.do_help("cd")
+        else:
+            self.current_directory = args
+            if args == "" or args == '~':
+                self.current_directory = '/'
 
-        execute_cmd('cd ' + args)
-        prompt.prompt = self.current_directory + '$'
+            execute_cmd('cd ' + args)
+            self.prompt = self.current_directory + ' > '
 
     def do_cat(self, args):
         """
@@ -108,7 +138,10 @@ class Shell(Cmd):
         Example:
             cat <file to display>
         """
-        execute_cmd('cd ' + self.current_directory + ' && cat ' + args)
+        if args == "":
+            self.do_help("cat")
+        else:
+            execute_cmd('cd ' + self.current_directory + ' && cat ' + args)
 
     def do_mkdir(self, args):
         """
@@ -116,7 +149,10 @@ class Shell(Cmd):
         Example:
             mkdir <new directory name>
         """
-        execute_cmd('mkdir ' + args)
+        if args == "":
+            self.do_help("mkdir")
+        else:
+            execute_cmd('mkdir ' + args)
 
     def do_rmdir(self, args):
         """
@@ -124,7 +160,10 @@ class Shell(Cmd):
         Example:
             rmdir <directory to remove name>
         """
-        execute_cmd('rmdir ' + args)
+        if args == "":
+            self.do_help("rmdir")
+        else:
+            execute_cmd('rmdir ' + args)
 
     def do_rm(self, args):
         """
@@ -132,7 +171,10 @@ class Shell(Cmd):
         Example:
             rm <filename>
         """
-        execute_cmd('rm ' + args)
+        if args == "":
+            self.do_help("rm")
+        else:
+            execute_cmd('rm ' + args)
 
     def do_clear(self, args):
         """
@@ -143,8 +185,38 @@ class Shell(Cmd):
         call('clear')
 
 
+class Shell(Cmd):
+
+    def do_shell(self, args):
+        """
+        This command opens up a shell to the Cornerstone Airlines Uptime Monitor Shell.
+        Example:
+            shell
+        """
+        newprompt = CornerstoneShell()
+        newprompt.prompt = newprompt.current_directory + '> '
+        newprompt.cmdloop()
+
+    def do_quit(self, args):
+        """
+            This command quits the interactive shell
+            Example:
+            quit
+        """
+        print("Goodbye!\n")
+        raise SystemExit
+
+    def do_clear(self, args):
+        """
+            This command clears the shell screen
+            Example:
+                clear
+        """
+        call('clear')
+
+
 if __name__ == '__main__':
     prompt = Shell()
-    prompt.prompt = '$'
+    prompt.prompt = '$ '
     prompt.cmdloop()
